gets mysql database schema fully wired up

+++ backend/pkg/sqlmanager/shared/types.go
@@ -15,10 +15,10 @@ type DatabaseSchemaRow struct {
 	DataType               string
 	ColumnDefault          string
 	IsNullable             string
-	CharacterMaximumLength int32
-	NumericPrecision       int32
-	NumericScale           int32
-	OrdinalPosition        int16
+	CharacterMaximumLength int
+	NumericPrecision       int
+	NumericScale           int
+	OrdinalPosition        int
 	GeneratedType          *string
 	IdentityGeneration     *string
 }
@@ -117,13 +117,13 @@ type TableConstraints struct {
 }
 
 type ColumnInfo struct {
-	OrdinalPosition        int32   // Specifies the sequence or order in which each column is defined within the table. Starts at 1 for the first column.
+	OrdinalPosition        int     // Specifies the sequence or order in which each column is defined within the table. Starts at 1 for the first column.
 	ColumnDefault          string  // Specifies the default value for a column, if any is set.
 	IsNullable             bool    // Specifies if the column is nullable or not.
 	DataType               string  // Specifies the data type of the column, i.e., bool, varchar, int, etc.
-	CharacterMaximumLength *int32  // Specifies the maximum allowable length of the column for character-based data types. For datatypes such as integers, boolean, dates etc. this is NULL.
-	NumericPrecision       *int32  // Specifies the precision for numeric data types. It represents the TOTAL count of significant digits in the whole number, that is, the number of digits to BOTH sides of the decimal point. Null for non-numeric data types.
-	NumericScale           *int32  // Specifies the scale of the column for numeric data types, specifically non-integers. It represents the number of digits to the RIGHT of the decimal point. Null for non-numeric data types and integers.
+	CharacterMaximumLength *int    // Specifies the maximum allowable length of the column for character-based data types. For datatypes such as integers, boolean, dates etc. this is NULL.
+	NumericPrecision       *int    // Specifies the precision for numeric data types. It represents the TOTAL count of significant digits in the whole number, that is, the number of digits to BOTH sides of the decimal point. Null for non-numeric data types.
+	NumericScale           *int    // Specifies the scale of the column for numeric data types, specifically non-integers. It represents the number of digits to the RIGHT of the decimal point. Null for non-numeric data types and integers.
 	IdentityGeneration     *string // Specifies the identity generation strategy for the column, if applicable.
 }
